+++ main.tf
@@ -1,10 +1,14 @@
 provider "alicloud" {
-  version                 = ">=1.64.0"
   profile                 = var.profile != "" ? var.profile : null
   shared_credentials_file = var.shared_credentials_file != "" ? var.shared_credentials_file : null
   region                  = var.region != "" ? var.region : null
   skip_region_validation  = var.skip_region_validation
   configuration_source    = "terraform-alicloud-modules/mongodb"
+}
+locals {
+  this_instance_id      = var.existing_instance_id != "" ? var.existing_instance_id : concat(alicloud_mongodb_instance.this.*.id, [""])[0]
+  create_more_resources = var.existing_instance_id != "" || var.create ? true : false
+  project               = "acs_mongodb"
 }
 
 resource "alicloud_mongodb_instance" "this" {
@@ -26,3 +30,98 @@ resource "alicloud_mongodb_instance" "this" {
   tags                 = var.tags
 }
 
+resource "alicloud_cms_alarm" "cpu_utilization" {
+  count   = local.create_more_resources && var.enable_alarm_rule ? 1 : 0
+  enabled = var.enable_alarm_rule
+  name    = var.alarm_rule_name
+  project = local.project
+  metric  = "CPUUtilization"
+  dimensions = {
+    instanceId = local.this_instance_id
+    device     = "/dev/vda1,/dev/vdb1"
+  }
+  statistics         = var.alarm_rule_statistics
+  period             = var.alarm_rule_period
+  operator           = var.alarm_rule_operator
+  threshold          = var.alarm_rule_threshold
+  triggered_count    = var.alarm_rule_triggered_count
+  contact_groups     = var.alarm_rule_contact_groups
+  silence_time       = var.alarm_rule_silence_time
+  effective_interval = var.alarm_rule_effective_interval
+}
+resource "alicloud_cms_alarm" "memory_utilization" {
+  count   = local.create_more_resources && var.enable_alarm_rule ? 1 : 0
+  enabled = var.enable_alarm_rule
+  name    = var.alarm_rule_name
+  project = local.project
+  metric  = "MemoryUtilization"
+  dimensions = {
+    instanceId = local.this_instance_id
+    device     = "/dev/vda1,/dev/vdb1"
+  }
+  statistics         = var.alarm_rule_statistics
+  period             = var.alarm_rule_period
+  operator           = var.alarm_rule_operator
+  threshold          = var.alarm_rule_threshold
+  triggered_count    = var.alarm_rule_triggered_count
+  contact_groups     = var.alarm_rule_contact_groups
+  silence_time       = var.alarm_rule_silence_time
+  effective_interval = var.alarm_rule_effective_interval
+}
+resource "alicloud_cms_alarm" "disk_utilization" {
+  count   = local.create_more_resources && var.enable_alarm_rule ? 1 : 0
+  enabled = var.enable_alarm_rule
+  name    = var.alarm_rule_name
+  project = local.project
+  metric  = "DiskUtilization"
+  dimensions = {
+    instanceId = local.this_instance_id
+    device     = "/dev/vda1,/dev/vdb1"
+  }
+  statistics         = var.alarm_rule_statistics
+  period             = var.alarm_rule_period
+  operator           = var.alarm_rule_operator
+  threshold          = var.alarm_rule_threshold
+  triggered_count    = var.alarm_rule_triggered_count
+  contact_groups     = var.alarm_rule_contact_groups
+  silence_time       = var.alarm_rule_silence_time
+  effective_interval = var.alarm_rule_effective_interval
+}
+resource "alicloud_cms_alarm" "intranet_in" {
+  count   = local.create_more_resources && var.enable_alarm_rule ? 1 : 0
+  enabled = var.enable_alarm_rule
+  name    = var.alarm_rule_name
+  project = local.project
+  metric  = "IntranetIn"
+  dimensions = {
+    instanceId = local.this_instance_id
+    device     = "/dev/vda1,/dev/vdb1"
+  }
+  statistics         = var.alarm_rule_statistics
+  period             = var.alarm_rule_period
+  operator           = var.alarm_rule_operator
+  threshold          = var.alarm_rule_threshold
+  triggered_count    = var.alarm_rule_triggered_count
+  contact_groups     = var.alarm_rule_contact_groups
+  silence_time       = var.alarm_rule_silence_time
+  effective_interval = var.alarm_rule_effective_interval
+}
+resource "alicloud_cms_alarm" "intranet_out" {
+  count   = local.create_more_resources && var.enable_alarm_rule ? 1 : 0
+  enabled = var.enable_alarm_rule
+  name    = var.alarm_rule_name
+  project = local.project
+  metric  = "IntranetOut"
+  dimensions = {
+    instanceId = local.this_instance_id
+    device     = "/dev/vda1,/dev/vdb1"
+  }
+  statistics         = var.alarm_rule_statistics
+  period             = var.alarm_rule_period
+  operator           = var.alarm_rule_operator
+  threshold          = var.alarm_rule_threshold
+  triggered_count    = var.alarm_rule_triggered_count
+  contact_groups     = var.alarm_rule_contact_groups
+  silence_time       = var.alarm_rule_silence_time
+  effective_interval = var.alarm_rule_effective_interval
+}